Add implementation comments for input and output
I have just realised that it would be possibly better if tasks had one
of three states (done, todo, in-progress) rather than two (done,
in progress). This will require some changes to the Task attribute class.

I would also like to add a taskID instance attribute to the Task class
and Task methods for updating, deleting, and modifying the state of a
Task (done, todo, in-progress).